+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+
+- Implemented `is_forestal` trait. Entities with this trait gain defense (+3) during combat when positioned on forest terrain. Also added forest terrain and forestal entities support to the CLI visualizer.
+
 ## [Unreleased]
 ### Added
 - Implemented `is_volcanic` trait. Entities with this trait thrive in extreme heat, gain energy when standing on `ash` terrain, and are immune to `fire` events.